implemented the dry principal
