+++ kolibri/plugins/learn/assets/src/views/LibraryPage/MoreNetworkDevices.vue
@@ -10,6 +10,7 @@
         :deviceName="device.device_name"
         :channels="device.channels"
         :allDevices="device"
+        :operatingSystem="device.operatingSystem"
       />
     </KGridItem>
   </KGrid>
