+++ kolibri/plugins/learn/assets/src/views/LibraryPage/PinnedNetworkResources.vue
@@ -1,13 +1,13 @@
 <template>
 
-  <KGrid>
+  <KGrid v-if="pinnedDevices !== null">
     <KGridItem
       v-for="device in pinnedDevices"
       :key="device.id"
     >
       <KGridItem>
         <h2>
-          <KIcon :icon="getDeviceIcon(device.operating_system)" />
+          <KIcon :icon="getDeviceIcon(device)" />
           <span class="device-name">{{ device.device_name }}</span>
         </h2>
       </KGridItem>
@@ -26,7 +26,6 @@
 
   import useKResponsiveWindow from 'kolibri.coreVue.composables.useKResponsiveWindow';
   import useDevices from './../../composables/useDevices';
-  import useChannels from './../../composables/useChannels';
   import ChannelCardGroupGrid from './../ChannelCardGroupGrid';
 
   export default {
@@ -37,29 +36,20 @@
     setup() {
       const { windowBreakpoint } = useKResponsiveWindow();
       const { baseurl, fetchDevices } = useDevices();
-      const { fetchChannels } = useChannels();
       return {
         windowBreakpoint,
-        fetchChannels,
         fetchDevices,
         baseurl,
       };
     },
-    data() {
-      return {
-        pinnedDevices: [],
-      };
+    props: {
+      pinnedDevices: {
+        type: Array,
+        required: true,
+      },
     },
-    created() {
-      this.fetchDevices().then(devices => {
-        const device = devices.filter(d => d.available);
-        device.forEach(element => {
-          this.fetchChannels({ baseurl: element.baseurl }).then(channel => {
-            this.$set(element, 'channels', channel);
-            this.pinnedDevices.push(element);
-          });
-        });
-      });
+    data() {
+      return {};
     },
     methods: {
       getDeviceIcon(device) {
